% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/career.long.data.R
\docType{data}
\name{career.long.data}
\alias{career.long.data}
\title{career.long.data: contains Career-long data are updated to end-of-2023}
\format{

}
\source{
Ioannidis, John P.A. (2024), “August 2024 data-update for "Updated science-wide author databases of standardized citation indicators"”, Elsevier Data Repository, V7, doi: 10.17632/btchxktzyw.7
}
\usage{
career.long.data
}
\description{
This dataset contains Table 1 (Table_1_Authors_career_2023_pubs_since_1788_wopp_extracted_202408.xlsx)of August 2024 data-update for "Updated science-wide author databases of standardized citation indicators"
}
\examples{
head(career.long.data)
}
\keyword{datasets}
