% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.year.data.R
\docType{data}
\name{single.year.data}
\alias{single.year.data}
\title{single.year.data: contains single recent year data pertain to citations received during calendar year 2023}
\format{

}
\source{
Ioannidis, John P.A. (2024), “August 2024 data-update for "Updated science-wide author databases of standardized citation indicators"”, Elsevier Data Repository, V7, doi: 10.17632/btchxktzyw.7
}
\usage{
single.year.data
}
\description{
This dataset contains Table 1 (Table_1_Authors_singleyr_2023_pubs_since_1788_wopp_extracted_202408.xlsx)of August 2024 data-update for "Updated science-wide author databases of standardized citation indicators"
}
\examples{
head(single.year.data)
}
\keyword{datasets}
